componenete para as duas telas de alterar senha
