+++ src/api/mock/data/chosen-question.json
@@ -1,0 +1,92 @@
+{
+  "data": {
+    "data": {
+      "customer": {
+        "id": "17",
+        "cpf": "...",
+        "rg": "",
+        "cnpj": null,
+        "companyName": null,
+        "stateScription": null,
+        "name": "...",
+        "welcome": "Jos\u00e9",
+        "phone": null,
+        "cellphone": "14991320957",
+        "birthDate": "",
+        "gender": null,
+        "email": "jo********tray.net.br",
+        "cards": null,
+        "hasCards": null,
+        "address": {
+          "id": "989",
+          "customerId": "17",
+          "type": "1",
+          "state": "SP",
+          "address": "Rua Ruy Mendes de Rosis",
+          "number": "**34",
+          "complement": null,
+          "neighborhood": "Jardim Infante Dom Henrique",
+          "zipCode": "17012-636",
+          "city": "Bauru",
+          "country": "Brazil",
+          "countryAbbr": "BRA",
+          "recipient": "Jos\u00e9 Augusto Lourencao",
+          "active": "1",
+          "description": "17012-636"
+        },
+        "billingAddress": null,
+        "addresses": [{
+          "id": "989",
+          "customerId": "17",
+          "type": "1",
+          "state": "SP",
+          "address": "Rua Ruy Mendes de Rosis",
+          "number": "**34",
+          "complement": null,
+          "neighborhood": "Jardim Infante Dom Henrique",
+          "zipCode": "17012-636",
+          "city": "Bauru",
+          "country": "Brazil",
+          "countryAbbr": "BRA",
+          "recipient": "Jos\u00e9 Augusto Lourencao",
+          "active": "1",
+          "description": "17012-636"
+        }],
+        "deliveryAddresses": [{
+          "id": "989",
+          "customerId": "17",
+          "type": "1",
+          "state": "SP",
+          "address": "Rua Ruy Mendes de Rosis",
+          "number": "**34",
+          "complement": null,
+          "neighborhood": "Jardim Infante Dom Henrique",
+          "zipCode": "17012-636",
+          "city": "Bauru",
+          "country": "Brazil",
+          "countryAbbr": "BRA",
+          "recipient": "Jos\u00e9 Augusto Lourencao",
+          "active": "1",
+          "description": "17012-636"
+        }],
+        "token": "5b30******************a03df8220b6102bdcdd2cda",
+        "loginMethod": "cpf",
+        "byPassword": false,
+        "attributes": [],
+        "facebookEmailMissing": false,
+        "credit": null,
+        "errors": [],
+        "checkTraycheckout": false,
+        "profile": null,
+        "hasEmailDomainMktplace": false,
+        "age": null,
+        "acceptNews": false,
+        "isDivergence": false,
+        "accounts": ["commerce"]
+      },
+      "message": true
+    },
+    "statusCode": 200
+  },
+  "statusCode": 200
+}
