+++ src/api/mock/data/error/chosen-question.json
@@ -1,0 +1,8 @@
+{
+  "data": {
+    "data": {
+      "message": false
+    },
+    "statusCode": 200
+  }
+}
