+++ src/api/mock/data/has-email-domain-mktplace.json
@@ -1,0 +1,11 @@
+{
+  "data":{
+      "data": {
+          "hasAccount": true,
+          "customer": {
+            "hasEmailDomainMktplace": false
+          }
+      }
+  },
+  "statusCode":200
+}
