+++ src/assets/icons/UserPassword.vue
@@ -1,0 +1,22 @@
+<template>
+  <figure class="tray-login__recover-password__figure">
+    <svg :width="134 * hasSize" :height="106 * hasSize" viewBox="0 0 134 106" fill="none" xmlns="http://www.w3.org/2000/svg">
+      <!-- eslint-disable-next-line -->
+      <path d="M46.9 53C52.2005 53 57.382 51.4458 61.7893 48.534C66.1965 45.6221 69.6315 41.4834 71.66 36.6411C73.6884 31.7989 74.2191 26.4706 73.185 21.3301C72.151 16.1896 69.5985 11.4678 65.8505 7.76168C62.1024 4.05559 57.3271 1.53171 52.1284 0.509202C46.9297 -0.513306 41.5411 0.0114819 36.6441 2.0172C31.747 4.02293 27.5614 7.4195 24.6166 11.7774C21.6718 16.1353 20.1 21.2588 20.1 26.5C20.1 33.5282 22.9236 40.2686 27.9495 45.2383C32.9755 50.208 39.7922 53 46.9 53ZM67 66.25C67.0151 64.0371 67.5987 61.8643 68.6959 59.9356C67.6909 59.832 66.7069 59.625 65.66 59.625H62.1634C57.3776 61.8073 52.1702 62.9374 46.9 62.9374C41.6298 62.9374 36.4224 61.8073 31.6366 59.625H28.14C20.6773 59.6267 13.5208 62.5587 8.24385 67.7766C2.96694 72.9944 0.00166522 80.0709 0 87.45L0 96.0625C0 98.6981 1.05884 101.226 2.94358 103.089C4.82832 104.953 7.38457 106 10.05 106H68.8634C67.6592 103.994 67.0161 101.708 67 99.375V66.25ZM127.3 59.625H120.6V43.0625C120.6 38.6699 118.835 34.4571 115.694 31.3511C112.553 28.245 108.292 26.5 103.85 26.5C99.4076 26.5 95.1472 28.245 92.006 31.3511C88.8647 34.4571 87.1 38.6699 87.1 43.0625V59.625H80.4C78.623 59.625 76.9189 60.323 75.6624 61.5654C74.4059 62.8078 73.7 64.4929 73.7 66.25V99.375C73.7 101.132 74.4059 102.817 75.6624 104.06C76.9189 105.302 78.623 106 80.4 106H127.3C129.077 106 130.781 105.302 132.038 104.06C133.294 102.817 134 101.132 134 99.375V66.25C134 64.4929 133.294 62.8078 132.038 61.5654C130.781 60.323 129.077 59.625 127.3 59.625ZM103.85 89.4375C102.525 89.4375 101.229 89.049 100.128 88.321C99.0259 87.593 98.1671 86.5583 97.66 85.3478C97.1529 84.1372 97.0202 82.8052 97.2787 81.52C97.5373 80.2349 98.1754 79.0544 99.1124 78.1279C100.049 77.2014 101.243 76.5704 102.543 76.3148C103.843 76.0592 105.19 76.1904 106.414 76.6918C107.638 77.1932 108.685 78.0424 109.421 79.1318C110.157 80.2213 110.55 81.5022 110.55 82.8125C110.55 84.5696 109.844 86.2547 108.588 87.4971C107.331 88.7395 105.627 89.4375 103.85 89.4375ZM110.55 59.625H97.15V43.0625C97.15 41.3054 97.8559 39.6204 99.1124 38.3779C100.369 37.1355 102.073 36.4375 103.85 36.4375C105.627 36.4375 107.331 37.1355 108.588 38.3779C109.844 39.6204 110.55 41.3054 110.55 43.0625V59.625Z" fill="#909090"/>
+    </svg>
+  </figure>
+</template>
+
+<script>
+export default {
+  props: ['size', 'identificationClasses'],
+  computed: {
+    /**
+     * Verifica se o parametro size tem valor
+     */
+    hasSize() {
+      return this.size ? this.size : 1;
+    },
+  },
+};
+</script>
